requirements.txt
Box2D==2.3.10
cffi==1.14.2
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
future==0.18.2
gym==0.18.0
kiwisolver==1.3.1
matplotlib==3.3.4
natsort==7.1.1
numpy==1.19.5
Pillow==7.2.0
pycparser==2.20
pyglet==1.5.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.1
PyYAML==5.4.1
scipy==1.5.4
six==1.15.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
typing-extensions==3.7.4.3
